Fix bug on security_level
The msg flags from a v3 message is defined as the last byte, for which one has to get the last 2 bits.

The error is that agents might use the remaining.

The fix here is to discard the left-outermost bits and keep the remaining two.

Thanks HoneyryderChuck for the explanation and fix